Replaced Repository with a class
